Restructure code and import hierarchy
