+++ galois/_codes/_reed_solomon.py
@@ -6,8 +6,9 @@ from numba import int64
 import numpy as np
 
 from .. import _lfsr
-from .._fields import Field, FieldClass, FieldArray, Poly, matlab_primitive_poly
+from .._fields import Field, FieldClass, FieldArray, matlab_primitive_poly
 from .._overrides import set_module
+from .._polys import Poly
 from .._prime import factors
 
 from ._cyclic import poly_to_generator_matrix, roots_to_parity_check_matrix
