+++ galois/_polys/__init__.py
@@ -1,0 +1,5 @@
+"""
+A subpackage containing arrays over Galois fields.
+"""
+from ._poly import *
+from ._poly_functions import *  # pylint: disable=redefined-builtin
