+++ galois/_polys/_poly_functions.py
@@ -5,9 +5,10 @@ from typing import Tuple, List
 
 import numpy as np
 
+from .._array import Array
 from .._overrides import set_module
 
-from ._main import FieldArray, Poly
+from ._poly import Poly
 
 __all__ = [
     "lagrange_poly",
@@ -108,7 +109,7 @@ def prod(*args):
 ###############################################################################
 
 @set_module("galois")
-def lagrange_poly(x: FieldArray, y: FieldArray) -> Poly:
+def lagrange_poly(x: Array, y: Array) -> Poly:
     r"""
     Computes the Lagrange interpolating polynomial :math:`L(x)` such that :math:`L(x_i) = y_i`.
 
@@ -164,9 +165,9 @@ def lagrange_poly(x: FieldArray, y: FieldArray) -> Poly:
 
         np.array_equal(L(x), y)
     """
-    if not isinstance(x, FieldArray):
+    if not isinstance(x, Array):
         raise TypeError(f"Argument `x` must be a Galois field array, not {type(x)}.")
-    if not isinstance(y, FieldArray):
+    if not isinstance(y, Array):
         raise TypeError(f"Argument `y` must be a Galois field array, not {type(y)}.")
     if not type(x) == type(y):  # pylint: disable=unidiomatic-typecheck
         raise TypeError(f"Arguments `x` and `y` must be over the same Galois field, not {type(x)} and {type(y)}.")
