+++ setup.cfg
@@ -43,7 +43,7 @@ project_urls =
     Twitter = https://twitter.com/galois_py
 
 [options]
-packages = galois, galois._codes, galois._databases, galois._fields
+packages = galois, galois._codes, galois._databases, galois._fields, galois._polys
 include_package_data = true
 test_suite = tests
 python_requires = >=3.6
